Internal checkout & plan cancel

+++ src/components/subscriptions/CurrentPlan.vue
@@ -1,22 +1,36 @@
 <template>
-  <div class="mt-4" v-show="isSubscribed">
-    <v-card>
-      <v-card-title class="cursive-font">Your current plan</v-card-title>
-      <v-card-title>{{ subscriptionState.activePlan.name }}</v-card-title>
-      <v-card-text>Started on : {{ planStart }}</v-card-text>
-      <v-card-text>Ends on : {{ planEnd }}</v-card-text>
+  <div class="d-flex flex-row mt-4" v-if="isSubscribed">
+    <v-card class="col-md-4 col-12" :loading="state.isCancelling" :disabled="state.isCancelling">
+      <v-card-title class="cursive-font">Your Subscription</v-card-title>
 
-      <v-card-text>
-        <v-btn @click="startPortalSession" :loading="state.isLoading" outlined color="primary">
-          Manage Plan
-        </v-btn>
+      <v-card-text class="d-flex flex-row">
+        Current plan:
+        <span class="ml-auto">{{ subscriptionState.activePlan.name }}</span>
       </v-card-text>
 
-      <a ref="portalLink" style="display: none" target="_blank"></a>
-    </v-card>
+      <v-card-text class="d-flex flex-row">
+        Started on :
+        <span class="ml-auto">{{ planStart }}</span>
+      </v-card-text>
 
-    <v-card v-show="!isSubscribed">
-      <v-card-text></v-card-text>
+      <v-card-text class="d-flex flex-row">
+        Ends on :
+        <span class="ml-auto">{{ planEnd }}</span>
+      </v-card-text>
+
+      <v-btn class="d-block mb-2" small elevation="0">
+        <b>
+          Change plan <v-icon>{{ icons.mdiChevronRight }}</v-icon>
+        </b>
+      </v-btn>
+
+      <v-btn @click="cancelSubscription" small elevation="0">
+        <b>
+          Cancel subscription <v-icon>{{ icons.mdiChevronRight }}</v-icon>
+        </b>
+      </v-btn>
+
+      <a ref="portalLink" style="display: none" target="_blank"></a>
     </v-card>
   </div>
 </template>
@@ -25,6 +39,8 @@
 import axios from "@axios";
 import { reactive, computed, ref } from "@vue/composition-api";
 import { useSubscription } from "@/composables/user/subscription";
+
+import { mdiChevronRight } from "@mdi/js";
 export default {
   name: "CurrentPlan",
   setup() {
@@ -35,20 +51,6 @@ export default {
       isCancelling: false,
       isLoading: false,
     });
-
-    async function startPortalSession() {
-      try {
-        state.isLoading = true;
-        const { data: session } = await axios.post("/subscriptions/portal");
-
-        portalLink.value.setAttribute("href", session.url);
-        portalLink.value.click();
-      } catch (err) {
-        console.log(err);
-      } finally {
-        state.isLoading = false;
-      }
-    }
 
     const planStart = computed(() => {
       const startDate = new Date(subscriptionState.activePlan.current_period_start * 1000);
@@ -63,12 +65,19 @@ export default {
     });
 
     async function cancelSubscription() {
+      const confirm = await this.$confirm(
+        "Are you sure you want to cancel your plan ? This action can not be undone",
+        { title: "warning" }
+      );
+
+      if (!confirm) return;
+
       try {
         state.isCancelling = true;
         await axios.delete("/subscriptions");
         setActivePlan({});
       } catch (err) {
-        console.log();
+        console.log(err);
       } finally {
         state.isCancelOpen = false;
         state.isCancelling = false;
@@ -85,7 +94,10 @@ export default {
       planEnd,
 
       cancelSubscription,
-      startPortalSession,
+
+      icons: {
+        mdiChevronRight,
+      },
     };
   },
 };
